+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - bb.pods: -> ed5e1f3
   - lsp4clj: 1.13.1 -> 2.0.1
 - Add selection to Extract Function code action #2118
+- Fix replace-refer-all mangling unrelated symbols #2150
 
 ## 2025.11.28-12.47.43